Added removal of Framework listeners in 4.3.9 which was mentioned already in other sections. As mentioned on osgi-dev by BJ
When a bundle is stopped, all the listeners registered with the framework are automatically removed. This should be discussed in 4.3.9 but I see that it is not :-( However 4.3.6 contains the following: "However, it is unnecessary to unregister services or Framework listeners, because they must be cleaned up by the Framework anyway." And 6.1.4.29 Bundle.stop also says: "Any listeners registered by this bundle must be removed." It would be nice if 4.3.9 also discussed this.