Remove getEntry methods and getHeader method. Rename BundleInfo to BundleRevision.
Removed getEntry methods and getHeader method 
since they only applied to the host bundle and
not to any attached fragments. There was no way to
access similar information from the attached 
fragments. We may added these methods to
BundleRevision later if we find a specific need.

BundleInfo was renamed to BundleRevision to reflect
that it represents a specific revision of a
bundle or fragment rathet than just some information
about a revision.