New build structure for hudson
Added a new project as head instead of osgi.ct. The osgi.build is now the head
to build, it depends on all other projects. This allows the osgi.ct to be build
without having the javadoc and other general tasks be done before. It was also
necessary to rename the files because the new bnd is using its defaults.