Move build output into separate repo
The build was placing the build output in cnf/repo
along side files which were committed to git. So
we needed a .gitignore file to ignore the build
output. It makes more sense to have a specific repo
which only holds the build output just like we have
a separate directory to hold .class files separate
from the .java source files.