Set o.o.framework.launch package version to 1.1
It was pointed out that we failed to increment the
package version of the framework.launch package
when we update the Bundle type in the framework 
package. Since Framework extends Bundle, changes 
to Bundle which necessitate the version increment
of the o.o.framework package also necessitate the
version increment of the o.o.framework.launch
package.