Bug 2184: Remove use of LogService in test case
The test case used log service to demonstrate the
web app could access the bundle context of the web
app bundle. However log service does not require
synchronous deliver to log readers so the test
could fail. All log service use is removed and
specific text is placed in the servlet result to
demonstrate access to the bundle context.