Add jsr14 flag to build
Specifying the jsr14 ant property to false will
cause the build to compiler all projects with 
target jsr14 twice. Once with target jsr14 and 
ee.minimum to make sure the API usage does not
exceed ee.minimum. The second compilation will be
target 1.5 and ee.j2se-1.5. The 1.5 class files 
will be used for the build output. Thus the shipped
companion code will work properly with Java 7 javac.

If the jsr14 ant property is not set or set to true
then there will only be a single compilation with
target jsr14.