Repository XML document schema
This is derived from the initial spec document but
with many corrections. However there is still a 
problem that needs to be resolved. The support for
lists is bad. I commented it out with a TODO:

Need to redo attribute values to support list. 
See for example blueprint.
(1) you can use angle brackets in the enumeration 
value.
(2) It is unclear how each scalar in the list is 
specified.