Adjust contentwidth to 5.0in
The source files specify contentwidth="6.5in", and that's what you get in the PDF, but you don't have that much space on the page.  Your page dimensions (as copied from the FrameMaker template) are:

7.44 inches paper width
.707 inches inner margin
.550 inches outer margin
1.0 inches body indent

That leaves 5.183 inches as the width of the body text area.  If you set the contentwidth="5.183in" in your imagedata elements, then the images will just fit the area width, if that is what you want.