Separate xi:include processing from FO styling
The build performed xi:include processing in the 
same step as FO styling. In preparation for change
markup support, we need to process xi:include in a
separate step from FO styling so we can process
the xi:include output for changes. We use a no-op
stylesheet to process the xi:include.

The build is also changed to allow the individual
steps of the book build to be done alone. This is
helpful for validating intermediate changes.

CatalogManager properties are also removed since
they don't seem to be needed or used.