Thread-safety: Don't re-read after null check to avoid returning null
The code attempted to use the racy single-check idiom (see Item 71 in
Effective Java, 2nd Ed.) but did not read the field into a local variable
before the check and possible return. The VM could reorder the two field
reads such that the read in the return statement could be done before
the read in the check. Thus the check could see non-null while the return
could see and return null. See also
http://jeremymanson.blogspot.co.uk/2008/12/benign-data-races-in-java.html