Hack RSA RI to not include osgi.enterprise and CM impl
The single bundle RSA RI is rather a mess of things. It included
osgi.enterprise-4.2 and exported all the packages. This created issues
with newer spec impls. It also included an old Felix CM impl which was
removed.

The single bundle idea is very hard to work with. We need to replace this
with a set of bundles like webcontainer. This allows easier analysis
and substitution with never versions of spec implementations.