osgi.specs: Add support for @ProviderType annotation
The @ProviderType annotation is used to mark the type to be implemented
by providers. It replaces the old @noimplement javadoc tag. If the
@ProviderType annotation is not present, then the @noimplement
javadoc tag will be used, if present.

The @ConsumerType annotation does not cause any text in the spec. It is
equivalent to the absence of @ProviderType (or @noimplement).