Add CT tests for RFC 188 Native Namespace
The existing native code tests have been updated
to test that the Bundle-NativeCode header is
properly transformed into an osgi.native
requirement.  Additionally the tests make sure 
the filter directive match expected attribute 
Maps.  For bundles that resolve the test asserts
that the wiring contains a proper wire to
a system osgi.native capability.

New tests were added to assert that a bundle
can specify an osgi.native requirement using
the Require-Capability header and that bundles
which attempt to specify osgi.native capabilities
with the Provide-Capability fail to install.