Bug 2543: VersionRange filter string must not match missing attribute
Since filter expressions do not support < and > operators, we must use
! on >= and <= operators. But using the ! operators means that false
evaluations due to missing attributes result in true. So when all the
expressions in the filter string use !, we must add a presence test to
the filter to ensure the attribute is present.

The CT for VersionRange is updated to verify that missing attributes
do not match the filter string.