book javadoc: sort Permissions subsections
The subsections of Permissions in the javadoc were not sorted and thus
the order could be random. This causes issues in the diff documents since
false changes appears. The xsl now sorts the subsections by name.

We also update the baselines to sort the subsections to avoid false
change reports.