osgi.extender: Remove the erroneous attribute names
These attributes are not defined in the spec (135.2). They are likely
the result of a copy-n-paste error when creating the ExtenderNamespace
class.

Since this package is not used at compile time, there are no binary
compatibility issues in deleting the constants. Any source code which
referenced the constants will need to be checked since the spec does
not define the attribute names.