Bug 2586: Add service hiding bundle to CT
The new org.osgi.test.filter bundle can be used to filter services out
of the CT environment. This can be necessary when there are multiple
instance of a service registered and we need to CT to test a specific
service. A specific instance of this is the LogService. OSGi has an RI
and the framework RI from Equinox also registers a LogService.