http RI: fix up dependencies
The RI jar file names must use the same version number as the
Bundle-Version in their manifest. Also, the RI is dependent on cm and
event api; not the CT. So we move the dependencies from the CT bnd file
into the RI lib file. Finally, the lib file is also renamed to 6.0 to
match the release (6) into which this change will go.