ds spec: Collapse FieldReference and LookupReference into Reference
At the 2014-12-18 CPEG call, we agreed to collapse the FieldReference
annotation in to the Reference annotation. This updated Reference
annotation can be applied to fields as well as methods. As part of this
change, I also collapsed the LookupReference annotation into the
Reference annotation so that we only have the one annotation for
references. This also affects the XML schema by removing the
field-reference element. The reference element can be used for event
strategy, lookup strategy and now field strategy.

Having a single annotation and a single xml element for references will
simplify the specification.