serviceobjects: Clarify unget behavior
A framework implementation is free to create more than one ServiceObjects
object for a given registered service. This is similar to
ServiceReference objects. However, like ServiceReference objects, all
the ServiceObjects objects for a given registered service are
interchangeable. That is, you can get from one ServiceObjects object and
successfully unget with another.