gradle: Use method names for bnd types
Some bnd types have non-private fields with the "same" name as the
getter method. For example: sourcepath and getSourcepath. Groovy
was picking the field instead of the getter. This caused an error
when the impl details of the field changed while the method did not.
So we now always use method calls in the Groovy code to make sure we
always call the getter and not access the field.