[CDI] Updated the DTOs
- Decoupled the template DTOs from the runtime state DTOs.
- Both can be matched by a name (id in the container case)
- The users can retrieve both by Bundle from the CdiRuntime service

The motivation is that names are needed anyway to make convenient the
use of template DTOs since one is not able to match the elements of
the static ContainerTemplateDTO to the "<runtime DTO>.template" values
without using deep equality. Once names are introduced it is cleaner
to only relay on them to match runtime do template elements.