ip ct: Fix manifest in ipa files
An interaction between bugs in Bnd and the IP RI caused tests to fail.
Bnd did not generate a manifest with a blank line after the main attrs
in the manifest and the IP RI requires as least one line after the
InitialProvisioning-Entries manifest entry to properly parse.

Since Bnd sorts the manifest entries, this fix ensures there is at least
one line in the manifest after the InitialProvisioning-Entries entry.