Clarify that getEntry method
Clarify that a path is used for getEntry and that the path used must not
start with a slash. This is important to framework implementors to
ensure they remove any leading slashes that may come from users of the
OSGi API for Bundle.getEntry or BundleWiring.findEntries methods which
do allow a leading slash.