filter: Update filter impl with caching of converted values
The values from the filter strings are parsed into strings. But they
often need to be converted to typed values for the actual comparison
with the input maps. This change caches the most recent conversion since
it is highly likely that the value types are highly consistent over the
various input maps. So we can save a lot of conversion of value to type.

This can be particularly beneficial for version attributes which are all
of type Version and we can be testing a specific version value across
many indexed resources of which can have many package capabilities.