spec: Change to control draft status of specs from bnd.bnd file
We can now configure the draft status of a specification from a
property in the osgi.specs/bnd.bnd file. This makes it easier
to build a proposed-final-draft or final spec since no docbook source
editing is required.

We now compute the draft.mode and draft.watermark.image XSLT parameters
from the bnd property and pass them to the XSLT processes. These
parameters control whether the generated specification's draft status.