promise: Mark internal callbacks so they can be called inline
When an internal inline callback is passed to onResolve of an already
resolved promise, we want to call the internal callback immediately on
the current thread. Internal inline callbacks do not call user code,
like a Function or Predicate, and are used to resolve the target
promise. Calling these internal callbacks inline can allow a newly
created DeferredPromiseImpl to be resolved and converted to a more
efficient resolved PromiseImpl before it is returned to the caller.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>