Introduce FeatureCapability and FeatureRequirement
The FeatureCapability and FeatureRequirement make the requirement and
capability API more conform other entities in the Feature model. They
also come with builders to make it easy to create them. Additionally
this extra layer allows for future extensibility.
FeatureCapability and FeatureRequirement extend the OSGi core
Capability and Requirement interfaces so they can be used as input to
APIs accepting those as well.