Remove Documented annotation from CLASS retention annotations
Documented has public API implications and, even though the annotations
are CLASS retention, causes java tools to expect them on the classpath.

This change results in a micro version bump to the packages and their
containing jar, so we must also bump the micro versions.

Includes some fixes to the new features and typedevent packages I made
while doing this larger change.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>