ci: Add step for release to create a staging repository id
We use a curl request to obtain the staging profile URI for the
org.osgi group id. And then we make a curl request to start a
staging repository. The URL to the created staging repository
is place in the GitHub actions env context for access by later
steps in the job.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>