[jaxrs] Improve the TCK compatibility with the JAX-RS specification
The JAX-RS SseEventSource is somewhat poorly specified, in that it will continually reconnect at the end of a stream, and is not required to notify of completion in this case. The server is expected to send a 204 response if the client is to stop connecting, at which point the client is notified that the stream has ended.

This change also adds an explicit message to set the client reconnect time to be one 'tick' which is 500 ms, ensuring that default delays in the client will not cause the test to time out.

Signed-off-by: Tim Ward <timothyjward@apache.org>