[jaxrs-tck] Remove race condition in async lifecycle test
The JAX-RS TCK has a race condition in the async lifecycle test - the test checks the value of *wasReleasedPreCompletion* without waiting for the service to be released. If the implementation releases the resource before committing the response (e.g. CXF) there is no issue, but if the response is committed eagerly before releasing the resource (e.g. Jersey) then the response may arrive before the resource is released. In the latter case we may get an incorrect failure. This can be easily fixed by adding a second semaphore marking the resource as released. This will occur rapidly after the result is received and only requires a short wait.

Signed-off-by: Tim Ward <timothyjward@apache.org>