[jakartars] Remove JakartarsEndpoint from the specification
The JakartarsEndpoint was intended to make it easy to find the URI for a named application or resource, and for that data to be automatically shared remotely. In reality this function was not used, or even implemented, by anyone. As a result this type will be removed from the specification, and can be re-added in future if it is actually needed by anyone.

Fixes #449

Signed-off-by: Tim Ward <timothyjward@apache.org>