core: Make InvalidSyntaxException a runtime exception
This is binary compatible since we are just introducing a new super
type whose super type is the previous super type.

We also remove "throws InvalidSyntaxException" from the framework API
since it is OSGi policy to not declare runtime exceptions in the throws
clause of a method's declaration. We still javadoc the throwing of the
exception.

Filter tests are updated to JUnit 5 and osgi-test.

Fixes https://github.com/osgi/osgi/issues/546

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>